+++ cli/src/react/__test__/ButtonArrowFunction.figma.tsx
@@ -1,5 +1,5 @@
 import figma from '../..'
 
-import { ButtonArrowFunction } from './TestComponents'
+import { ButtonArrowFunction } from './components/TestComponents'
 
 figma.connect(ButtonArrowFunction, 'ui/button')