+++ cli/src/react/__test__/DefaultImport.figma.tsx
@@ -1,4 +1,4 @@
 import figma from '../..'
-import RenamedButton from './TestComponents'
+import RenamedButton from './components/TestComponents'
 
 figma.connect(RenamedButton, 'ui/button')