+++ cli/src/react/__test__/ForwardRefComponent.figma.tsx
@@ -1,4 +1,4 @@
 import figma from '../..'
-import { ForwardRefButton } from './TestComponents'
+import { ForwardRefButton } from './components/TestComponents'
 
 figma.connect(ForwardRefButton, 'ui/button')