+++ cli/src/react/__test__/MemoizedComponent.figma.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import figma from '../..'
-import { MemoButton } from './TestComponents'
+import { MemoButton } from './components/TestComponents'
 
 figma.connect(MemoButton, 'ui/button', {
   example: () => <MemoButton>Click me</MemoButton>,