+++ cli/src/react/__test__/NamespacedComponent.figma.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
 import figma from '../..'
-import { NamespacedComponents } from './TestComponents'
+import { NamespacedComponents } from './components/TestComponents'
 
 figma.connect(NamespacedComponents.Button, 'ui/button', {
   example: () => <NamespacedComponents.Button />,