+++ cli/src/react/__test__/NoProps.figma.tsx
@@ -1,4 +1,4 @@
 import figma from '../..'
-import { ComponentWithoutProps } from './TestComponents'
+import { ComponentWithoutProps } from './components/TestComponents'
 
 figma.connect(ComponentWithoutProps, 'ui/button')