+++ cli/src/react/__test__/PropMappings.figma.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import figma from '../..'
-import { Button } from './TestComponents'
+import { Button } from './components/TestComponents'
 
 figma.connect(Button, 'propsInline', {
   props: {