+++ cli/src/react/__test__/TopLevelComponent.figma.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import figma from '../..'
-import { Button } from './components/TestComponents'
+import { Button } from './TestTopLevelComponent'
 
 figma.connect(Button, 'ui/button', {
   example: () => <Button>Click me</Button>,