+++ cli/src/react/__test__/VariableRefFigmaNode.figma.tsx
@@ -1,5 +1,5 @@
 import figma from '../..'
-import Button from './TestComponents'
+import Button from './components/TestComponents'
 
 const buttonURL = 'identifierAsUrl'