+++ swiftui/lib/ParserTypes/FigmaChildren.swift
@@ -3,13 +3,13 @@ import Foundation
 
 public struct FigmaChildren: Encodable, Equatable {
     struct Args: Encodable, Equatable {
-        let layerNames: [String]
+        let layers: [String]
     }
     let kind = "children"
     let args: Args
     
-    init(layerNames: [String]) {
-        self.args = Args(layerNames: layerNames)
+    init(layers: [String]) {
+        self.args = Args(layers: layers)
     }
 }